api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
clickhouse-connect==0.6.13
SQLAlchemy==2.0.23
psycopg[binary]==3.1.18
python-dotenv==1.0.1
alembic==1.12.0
pydantic==2.9.2
httpx==0.27.0
